feat: 全链路切换为严格 FP32
- 统一训练、EMA、checkpoint、publication 和推理精度为 FP32
- 禁用 BF16、FP16 autocast 及 TF32
- 模型产物与配置升级至 schema v7
- 拒绝非 FP32 输入、权重及旧版 v6 产物
- 重构分阶段 GPU smoke 并修复显存测量问题
- 补充长期 FP32 策略、选择依据及实测文档

+++ src/zero_ttt/inference/batching.py
@@ -164,7 +164,7 @@ class BatchedInferenceBroker:
         output = self.evaluator.evaluate(batch)
         elapsed = time.perf_counter() - started
         logits = output.policy_logits.detach().cpu().numpy()
-        values = output.value.detach().float().reshape(len(unique)).cpu().numpy()
+        values = output.value.detach().reshape(len(unique)).cpu().numpy()
         ownership = self._optional_numpy(output.ownership)
         scores = self._optional_numpy(output.score_margin, reshape=len(unique))
         if len(logits) != len(unique):
@@ -187,7 +187,7 @@ class BatchedInferenceBroker:
     def _optional_numpy(tensor, *, reshape: int | None = None):
         if tensor is None:
             return None
-        value = tensor.detach().float()
+        value = tensor.detach()
         if reshape is not None:
             value = value.reshape(reshape)
         return value.cpu().numpy()
